Extend value_set_of_closures to have less fields for classic mode.
A variant is added to differentiate between fields that are only
required in normal mode (when flambda peforms aggresive optimization) and
classic mode (when flambda doesn't optimize function calls). The file
changed is in [middle_ends/simple_value_approx.mli] - other files that uses
the type are changed so that the program type checks.