Flambda: Add [Simple_value_approx.function_declarations]
Adds a type to represent an approximation of function_declarations.
This type is semantically different from [Flambda.function_declarations].
This change would permit to further modification to its fields in an
effort to reduce the size of cmx files in Flambda.

This change on its own does not compile, but encompasses a large enough
modification to be a commit on its own.