Flambda: Add is_classic_mode field to value_set_of_closures.
The field provides an invariance on the contents of function bodies of the
underlying function declarations. In -Oclassic, it is not necessary to store
bodies in the cmx files (as most of the inlining decision is made at
definition site).

This commit also fixes [A.create_classic_function_declarations] which
previously incorrectly throws away all the given function body. The
correct behaviour in classic mode is to keep the body selectively.